Moving app/building to component/building, adding pages to different button in headers
