+++ .env
@@ -1,4 +1,4 @@
 NEXT_PUBLIC_CLERK_SIGN_IN_URL='/sign-in'
 NEXT_PUBLIC_CLERK_SIGN_UP_URL='/sign-up'
-NEXT_PUBLIC_CLERK_AFTER_SIGN_IN_URL='/dashboard'
-NEXT_PUBLIC_CLERK_AFTER_SIGN_UP_URL='/dashboard'
+NEXT_PUBLIC_CLERK_AFTER_SIGN_IN_URL='/'
+NEXT_PUBLIC_CLERK_AFTER_SIGN_UP_URL='/'
