+++ app/about-us/page.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function AboutUs() {
+  return (
+    <div>AboutUs</div>
+  )
+}
