+++ app/components/buildings/NGE.module.css
