--- app/homepage/AboutUs.jsx
@@ -1,17 +1,0 @@
-
-import React from 'react';
-import Header  from '../components/Header'; 
-
-
-const AboutUs = () => {
-  return (
-    <div>
-      <Header />
-      <h1>About Us</h1>
-      <p>HDGAHDGHAGDHAGDHAG</p>
-      {/* Add more content as needed */}
-    </div>
-  );
-};
-
-export default AboutUs;
