+++ app/page.jsx
@@ -1,14 +1,10 @@
-import Image from "next/image";
-import { UserButton } from "@clerk/nextjs";
-import NGE from "/app/buildings/NGE.jsx";
-import AboutUs from "./homepage/AboutUs";
 
-
+// TODO add log out button for the user
 
 export default function Home() {
   return (
     <div>
-      <NGE/>
+      Home
     </div>
   );
 }
