+++ app/room-details/page.jsx
@@ -1,0 +1,8 @@
+import React from 'react'
+import NGE from "/app/components/buildings/NGE.jsx";
+
+export default function RoomDetails() {
+  return (
+    <div><NGE/></div>
+  )
+}
