+++ app/search-room/page.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function SearchRoom() {
+  return (
+    <div>SearchRoom</div>
+  )
+}
